extract gauss
