+++ scanpy/neighbors/common.py
@@ -1,0 +1,27 @@
+import numpy as np
+from scipy.sparse import csr_matrix
+
+
+def _get_sparse_matrix_from_indices_distances_numpy(
+    indices, distances, n_obs, n_neighbors
+):
+    n_nonzero = n_obs * n_neighbors
+    indptr = np.arange(0, n_nonzero + 1, n_neighbors)
+    D = csr_matrix(
+        (
+            distances.copy().ravel(),  # copy the data, otherwise strange behavior here
+            indices.copy().ravel(),
+            indptr,
+        ),
+        shape=(n_obs, n_obs),
+    )
+    D.eliminate_zeros()
+    return D
+
+
+def _get_indices_distances_from_dense_matrix(D, n_neighbors: int):
+    sample_range = np.arange(D.shape[0])[:, None]
+    indices = np.argpartition(D, n_neighbors - 1, axis=1)[:, :n_neighbors]
+    indices = indices[sample_range, np.argsort(D[sample_range, indices])]
+    distances = D[sample_range, indices]
+    return indices, distances
